Adding truncate feature to load_dimension.py

+++ plugins/final_project_operators/load_dimension.py
@@ -11,18 +11,21 @@ class LoadDimensionOperator(BaseOperator):
                  redshift_conn_id,
                  sql,
                  table,
+                 truncate=True,
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
         self.redshift_conn_id = redshift_conn_id
         self.sql = sql
         self.table = table
+        self.truncate = truncate
 
     def execute(self, context):
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
 
-        self.log.info("Clearing data from destination dimension table {}".format(self.table))
-        redshift.run("DELETE FROM {}".format(self.table))
+        if self.truncate:
+            self.log.info("Clearing data from destination dimension table {}".format(self.table))
+            redshift.run("DELETE FROM {}".format(self.table))
 
         self.log.info('Inserting into dimension table {}'.format(self.table))
         redshift.run('INSERT INTO {} {}'.format(self.table, self.sql))
